[KT] Get rid of duplicates from tests
