+++ test/kt/esimd_radix_sort.cpp
@@ -8,6 +8,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "../support/test_config.h"
+#include "esimd_radix_sort_utils.h"
 
 #if TEST_DPCPP_BACKEND_PRESENT
 
@@ -32,28 +33,12 @@
 
 #include <vector>
 #include <algorithm>
-#include <random>
 #include <string>
+#include <cstdint>
+
+#if LOG_TEST_INFO
 #include <iostream>
-#include <cmath>
-#include <limits>
-
-#ifndef LOG_TEST_INFO
-#    define LOG_TEST_INFO 0
 #endif
-
-template <typename T, bool Order>
-struct Compare : public std::less<T>
-{
-};
-
-template <typename T>
-struct Compare<T, false> : public std::greater<T>
-{
-};
-
-constexpr bool Ascending = true;
-constexpr bool Descending = false;
 
 #ifndef TEST_DATA_TYPE
 #    define TEST_DATA_TYPE int
@@ -67,212 +52,8 @@ constexpr bool Descending = false;
 #    define TEST_WORK_GROUP_SIZE 64
 #endif
 
-constexpr std::uint8_t RadixBits = 8;
-
 using ParamType = oneapi::dpl::experimental::kt::kernel_param<TEST_DATA_PER_WORK_ITEM, TEST_WORK_GROUP_SIZE>;
 constexpr ParamType kernel_parameters;
-
-#if LOG_TEST_INFO
-struct TypeInfo
-{
-    template <typename T>
-    const std::string&
-    name()
-    {
-        static const std::string TypeName = "unknown type name";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<char>()
-    {
-        static const std::string TypeName = "char";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<int8_t>()
-    {
-        static const std::string TypeName = "int8_t";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<uint8_t>()
-    {
-        static const std::string TypeName = "uint8_t";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<int16_t>()
-    {
-        static const std::string TypeName = "int16_t";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<uint16_t>()
-    {
-        static const std::string TypeName = "uint16_t";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<uint32_t>()
-    {
-        static const std::string TypeName = "uint32_t";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<uint64_t>()
-    {
-        static const std::string TypeName = "uint64_t";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<int64_t>()
-    {
-        static const std::string TypeName = "int64_t";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<int>()
-    {
-        static const std::string TypeName = "int";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<float>()
-    {
-        static const std::string TypeName = "float";
-        return TypeName;
-    }
-
-    template <>
-    const std::string&
-    name<double>()
-    {
-        static const std::string TypeName = "double";
-        return TypeName;
-    }
-};
-
-struct USMAllocPresentation
-{
-    template <sycl::usm::alloc>
-    const std::string&
-    name()
-    {
-        static const std::string USMAllocTypeName = "unknown";
-        return USMAllocTypeName;
-    }
-
-    template <>
-    const std::string&
-    name<sycl::usm::alloc::host>()
-    {
-        static const std::string USMAllocTypeName = "sycl::usm::alloc::host";
-        return USMAllocTypeName;
-    }
-
-    template <>
-    const std::string&
-    name<sycl::usm::alloc::device>()
-    {
-        static const std::string USMAllocTypeName = "sycl::usm::alloc::device";
-        return USMAllocTypeName;
-    }
-
-    template <>
-    const std::string&
-    name<sycl::usm::alloc::shared>()
-    {
-        static const std::string USMAllocTypeName = "sycl::usm::alloc::shared";
-        return USMAllocTypeName;
-    }
-
-    template <>
-    const std::string&
-    name<sycl::usm::alloc::unknown>()
-    {
-        static const std::string USMAllocTypeName = "sycl::usm::alloc::unknown";
-        return USMAllocTypeName;
-    }
-};
-#endif // LOG_TEST_INFO
-
-template <typename T>
-typename ::std::enable_if_t<std::is_arithmetic_v<T>, void>
-generate_data(T* input, std::size_t size)
-{
-    std::default_random_engine gen{42};
-    std::size_t unique_threshold = 75 * size / 100;
-    if constexpr (sizeof(T) < sizeof(short)) // no uniform_int_distribution for chars
-    {
-        std::uniform_int_distribution<int> dist(std::numeric_limits<T>::lowest(), std::numeric_limits<T>::max());
-        std::generate(input, input + unique_threshold, [&] { return T(dist(gen)); });
-    }
-    else if constexpr (std::is_integral_v<T>)
-    {
-        std::uniform_int_distribution<T> dist(std::numeric_limits<T>::lowest(), std::numeric_limits<T>::max());
-        std::generate(input, input + unique_threshold, [&] { return dist(gen); });
-    }
-    else
-    {
-        std::uniform_real_distribution<T> dist_real(std::numeric_limits<T>::min(), log2(1e12));
-        std::uniform_int_distribution<int> dist_binary(0, 1);
-        auto randomly_signed_real = [&dist_real, &dist_binary, &gen]()
-        {
-            auto v = exp2(dist_real(gen));
-            return dist_binary(gen) == 0 ? v : -v;
-        };
-        std::generate(input, input + unique_threshold, [&] { return randomly_signed_real(); });
-    }
-    for (uint32_t i = 0, j = unique_threshold; j < size; ++i, ++j)
-    {
-        input[j] = input[i];
-    }
-}
-
-template <typename Container1, typename Container2>
-void
-print_data(const Container1& expected, const Container2& actual, std::size_t first, std::size_t n = 0)
-{
-    if (expected.size() <= first)
-        return;
-    if (n == 0 || expected.size() < first + n)
-        n = expected.size() - first;
-
-    if constexpr (std::is_floating_point_v<typename Container1::value_type>)
-        std::cout << std::hexfloat;
-    else
-        std::cout << std::hex;
-
-    for (std::size_t i = first; i < first + n; ++i)
-    {
-        std::cout << actual[i] << " --- " << expected[i] << std::endl;
-    }
-
-    if constexpr (std::is_floating_point_v<typename Container1::value_type>)
-        std::cout << std::defaultfloat << std::endl;
-    else
-        std::cout << std::dec << std::endl;
-}
 
 #if _ENABLE_RANGES_TESTING
 template <typename T, bool IsAscending, std::uint8_t RadixBits, typename KernelParam>
@@ -283,7 +64,7 @@ test_all_view(sycl::queue q, std::size_t size, KernelParam param)
     std::cout << "\ttest_all_view(" << size << ") : " << TypeInfo().name<T>() << std::endl;
 #endif
     std::vector<T> input(size);
-    generate_data(input.data(), size);
+    generate_data(input.data(), size, 42);
     std::vector<T> ref(input);
     std::stable_sort(std::begin(ref), std::end(ref), Compare<T, IsAscending>{});
     {
@@ -305,7 +86,7 @@ test_subrange_view(sycl::queue q, std::size_t size, KernelParam param)
               << std::endl;
 #endif
     std::vector<T> expected(size);
-    generate_data(expected.data(), size);
+    generate_data(expected.data(), size, 42);
 
     TestUtils::usm_data_transfer<sycl::usm::alloc::device, T> dt_input(q, expected.begin(), expected.end());
 
@@ -332,7 +113,7 @@ test_usm(sycl::queue q, std::size_t size, KernelParam param)
               << IsAscending << ">(" << size << ");" << std::endl;
 #endif
     std::vector<T> expected(size);
-    generate_data(expected.data(), size);
+    generate_data(expected.data(), size, 42);
 
     TestUtils::usm_data_transfer<_alloc_type, T> dt_input(q, expected.begin(), expected.end());
 
@@ -357,7 +138,7 @@ test_sycl_iterators(sycl::queue q, std::size_t size, KernelParam param)
     std::cout << "\t\ttest_sycl_iterators<" << TypeInfo().name<T>() << ">(" << size << ");" << std::endl;
 #endif
     std::vector<T> input(size);
-    generate_data(input.data(), size);
+    generate_data(input.data(), size, 42);
     std::vector<T> ref(input);
     std::stable_sort(std::begin(ref), std::end(ref), Compare<T, IsAscending>{});
     {
@@ -431,15 +212,15 @@ main()
 #if TEST_LONG_RUN
             for (auto size : sizes)
             {
-                test_general_cases<TEST_DATA_TYPE, Ascending, RadixBits>(q, size, kernel_parameters);
-                test_general_cases<TEST_DATA_TYPE, Descending, RadixBits>(q, size, kernel_parameters);
+                test_general_cases<TEST_DATA_TYPE, Ascending, TestRadixBits>(q, size, kernel_parameters);
+                test_general_cases<TEST_DATA_TYPE, Descending, TestRadixBits>(q, size, kernel_parameters);
             }
-            test_small_sizes<TEST_DATA_TYPE, Ascending, RadixBits>(q, kernel_parameters);
+            test_small_sizes<TEST_DATA_TYPE, Ascending, TestRadixBits>(q, kernel_parameters);
 #else
             for (auto size : sizes)
             {
-                test_usm<TEST_DATA_TYPE, Ascending, RadixBits, sycl::usm::alloc::shared>(q, size, kernel_parameters);
-                test_usm<TEST_DATA_TYPE, Descending, RadixBits, sycl::usm::alloc::shared>(q, size, kernel_parameters);
+                test_usm<TEST_DATA_TYPE, Ascending, TestRadixBits, sycl::usm::alloc::shared>(q, size, kernel_parameters);
+                test_usm<TEST_DATA_TYPE, Descending, TestRadixBits, sycl::usm::alloc::shared>(q, size, kernel_parameters);
             }
 #endif // TEST_LONG_RUN
         }
